<fix> 실험용

+++ Assets/RawData/Scripts/Skill/PlayerSkillInfo.cs
@@ -18,7 +18,7 @@ public class PlayerSkillInfo
     public byte _upgrade;
     public string _prefabPath;
     public string _spritePath;
-    public int _cost;
+    public float _cost;
 
     private Sprite _sprite;
